Rename some commands for examples to follow naming convention
