+++ examples/serverinfo/include/precondition_require_guild.h
