+++ examples/serverinfo/precondition_require_guild.cpp
